+++ apps/api/src/app/core/passport.ts
@@ -3,19 +3,20 @@ import { environment as env } from '@environments/environment';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
 import { User } from '@models';
+import { IUser } from '@csl/shared';
 
 export function setupPassport() {
-  // Create user session
-  passport.serializeUser((user: any, done: any) => {
-    done(null, user.id);
-  });
+	// Create user session
+	passport.serializeUser((user: IUser, done) => {
+		done(null, user.id);
+	});
 
-  // Decrypt session and find user
-  passport.deserializeUser((id, done) => {
-    User.findOne({ id: id }).then((user) => {
-      done(null, user);
-    });
-  });
+	// Decrypt session and find user
+	passport.deserializeUser((id: string, done) => {
+		User.findOne({ id }).then((user: IUser) => {
+			done(null, user);
+		});
+	});
 
   // Google Login
   passport.use(
@@ -25,34 +26,34 @@ export function setupPassport() {
         clientSecret: env.GOOGLE_CLIENT_SECRET,
         callbackURL: `${env.api}/auth/redirect`,
       },
-      (accessToken: any, refreshToken: any, profile: any, done: any) => {
+      (accessToken: string, refreshToken: string, profile, done) => {
         const id = profile.id;
         const photoURL = profile.photos[0].value;
         const email = profile.emails[0].value;
 
-        User.findOne({ id }).then(async (user) => {
-          if (user) {
-            await user.updateOne({ photoURL });
+				User.findOne({ id }).then(async (user) => {
+					if (user) {
+						await user.updateOne({ photoURL });
 
-            done(null, user);
-          } else {
-            User.findOne({ email }).then(async (user) => {
-              if (user) {
-                await user.updateOne({
-                  id,
-                  photoURL,
-                });
+						done(null, user);
+					} else {
+						User.findOne({ email }).then(async (user) => {
+							if (user) {
+								await user.updateOne({
+									id,
+									photoURL,
+								});
 
-                User.findOne({ id }).then((user) => {
-                  done(null, user);
-                });
-              } else {
-                done(null, null);
-              }
-            });
-          }
-        });
-      }
-    )
-  );
+								User.findOne({ id }).then((user) => {
+									done(null, user);
+								});
+							} else {
+								done(null, null);
+							}
+						});
+					}
+				});
+			}
+		)
+	);
 }