+++ apps/client/src/app/admin/admin-routing.module.ts
@@ -18,6 +18,7 @@ import { ClassiComponent } from '@shared/components/classi/classi.component';
 import { SingleClassComponent } from '@shared/components/single-class/single-class.component';
 import { CsvComponent } from '@shared/components/csv/csv.component';
 import { CommissioniComponent } from './components/commissioni/commissioni.component';
+import { ServiceAccountView } from './views/service-account/service-account.view';
 
 const routes: Routes = [
 	{
@@ -40,6 +41,7 @@ const routes: Routes = [
 			{ path: 'events', component: LogsComponent },
 			{ path: 'coge', component: CogeComponent },
 			{ path: 'coge/create', component: CreateCourseComponent },
+			{ path: 'service', component: ServiceAccountView },
 		],
 	},
 ];