+++ apps/maintenance/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [1.8.33](https://github.com/LuckeeDev/csl/compare/v1.8.31...v1.8.33) (2021-03-25)
+## [1.9.1](https://github.com/LuckeeDev/csl/compare/v1.9.0...v1.9.1) (2021-03-25)
+
+
+
+# [1.9.0](https://github.com/LuckeeDev/csl/compare/v1.8.32...v1.9.0) (2021-03-25)